dashboard and graphical interface idea
builds an example dashboard with asynchronously loading components.
kind of a hack as i'm just playing with some ideas, but it's a hack that
could be useful as some sort of yams.host/dashboards/example.

the actual work to do user-defined field matching ended up being kind of
non-trivial (esp for n:depth:json:*:status globbing).  this also
introduces a JSON syntax for retreiving fields, which i'm not sure that
i'm in love with.